# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Production-Grade Web3 RFQ Market Maker API
  version: 0.1.0
  description: API draft for RFQ quote, submit, administrative quote control, health, and metrics flows. Every response includes an x-trace-id header for request correlation; structured error bodies also include the same traceId value. Clients may send a safe tr_-prefixed x-trace-id request header, which the gateway echoes after length and character validation; unsafe values fall back to a generated request id. Business and administrative routes require a scoped x-api-key in non-local deployments; only the key id and SHA-256 secret digest are stored server-side. Quotes belong to the authenticated stable principal, not an individual key id, so same-principal key rotation preserves access while cross-principal quote, settlement, and hedge lookups return the same 404 response as an absent resource; submit and PnL are scoped identically. The principal is off-chain authorization metadata and is never part of a public response or EIP-712 payload. Browser access is controlled by the RFQ_CORS_ALLOWED_ORIGINS allowlist, which only accepts normalized HTTP(S) URL origins without path, query, fragment, credentials, or wildcards. The gateway emits no-store cache control and baseline browser security headers on every response. Unknown routes and unsupported methods return the standard ErrorResponse shape.
servers:
  - url: http://localhost:3000
security:
  - ApiKeyAuth: []
paths:
  /quote:
    post:
      operationId: createQuote
      summary: Request an RFQ quote
      description: >
        Prices and risk-checks the request before signing. The shared administrative
        global quote-control state is checked first, followed by the normalized chain
        and token-pair state after request validation; either pause or an unreadable
        state fails closed without invalidating signed quotes already issued. Receipt-confirmed runtimes
        also read the configured settlement contract treasury and tokenOut balance at
        one chain block, then atomically reserve unexpired output capacity across API
        replicas. Insufficient or unavailable Treasury liquidity fails closed before
        the EIP-712 signer is called.
      x-required-scope: quote:write
      tags:
        - Quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteRequest"
            examples:
              default:
                value:
                  chainId: 1
                  user: "0x0000000000000000000000000000000000000001"
                  tokenIn: "0x0000000000000000000000000000000000000002"
                  tokenOut: "0x0000000000000000000000000000000000000003"
                  amountIn: "1000000000"
                  slippageBps: 50
      responses:
        "200":
          description: Signed quote
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteResponse"
        "401":
          description: Missing, malformed, invalid, or expired API key
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: API key lacks the scope required by this operation
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                default:
                  value:
                    quoteId: "q_1893456000000_1"
                    snapshotId: "snapshot_1_00000000_00000000"
                    amountOut: "998400000"
                    minAmountOut: "993408000"
                    deadline: 1893456030
                    nonce: "1893456000000001"
                    signature: "0x1111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111"
        "400":
          description: Invalid request
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Risk rejected
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Request body too large
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
            Retry-After:
              description: Seconds until the client should retry.
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Quote creation paused, quote-control or rate-limit store unavailable, market data unavailable, stale, too far in the future, invalid, routing unavailable, quote store unavailable, signer unavailable, or another quote dependency unavailable
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /submit:
    post:
      operationId: submitQuote
      summary: Confirm a wallet transaction or run an explicitly enabled local simulation
      x-required-scope: submit:write
      description: >
        In receipt-confirmed mode the client supplies the transaction hash produced by
        its wallet. The backend independently waits for configured confirmations,
        verifies the successful receipt, transaction sender, target settlement contract,
        and decoded submitQuote calldata, then requires exactly one matching QuoteSettled event before applying
        inventory, hedge, PnL, metrics, or quote-status side effects. Synthetic settlement
        is available only when explicitly enabled for local reference workflows.
        Hedge intent creation failure does not roll back an accepted settlement;
        the response remains accepted but omits hedgeOrderId and the metrics endpoint
        increments rfq_hedge_intent_errors_total. The reference backend also records
        a bounded output-token quote risk penalty so follow-up quotes become more
        conservative after post-trade hedge failures. RFQSettlement QuoteSettled events
        are the final settlement source of truth.
        Submit first validates the canonical request shape, resolves the stored signed
        quote by chainId/user/nonce, and rejects signatures that differ from the
        signature returned by /quote before signer recovery or settlement verification.
      tags:
        - Submit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitQuoteRequest"
      responses:
        "202":
          description: Submit accepted for processing
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitQuoteResponse"
        "401":
          description: Missing, malformed, invalid, or expired API key
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: API key lacks the scope required by this operation
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Invalid submit request
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Signed quote not found
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Quote expired, already used, signature invalid, or rejected by settlement verification
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Request body too large
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
            Retry-After:
              description: Seconds until the client should retry.
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Submit reservation store, rate limit store, settlement verifier, chain RPC, or another settlement dependency is unavailable before settlement is applied
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /quote/{quoteId}:
    get:
      operationId: getQuote
      summary: Get quote status by id
      x-required-scope: status:read
      tags:
        - Quote
      parameters:
        - name: quoteId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: "^[A-Za-z0-9_:-]+$"
          description: Quote identifier returned by POST /quote. The gateway accepts only 1 to 128 letters, numbers, underscore, colon, or hyphen characters.
      responses:
        "200":
          description: Quote status
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteStatus"
        "401":
          description: Missing, malformed, invalid, or expired API key
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: API key lacks the scope required by this operation
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Quote not found
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
            Retry-After:
              description: Seconds until the client should retry.
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Rate limit store or quote status store unavailable
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /hedges/{hedgeOrderId}:
    get:
      operationId: getHedgeIntent
      summary: Get hedge intent status by id
      x-required-scope: status:read
      description: >
        Returns the hedge intent created after a submit request updates
        inventory, including its current queued, filled, or failed lifecycle
        status and terminal venue reconciliation metadata when available.
        Non-local deployments persist this state in PostgreSQL using one
        idempotent hedge intent per settlement event. A separate lease-based worker
        queries the configured CEX by deterministic client order id before submission
        and persists only explicit terminal venue outcomes.
      tags:
        - Hedge
      parameters:
        - name: hedgeOrderId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: "^[A-Za-z0-9_:-]+$"
          description: Hedge order identifier returned by POST /submit. The gateway accepts only 1 to 128 letters, numbers, underscore, colon, or hyphen characters.
      responses:
        "200":
          description: Hedge intent status
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HedgeIntentStatus"
        "401":
          description: Missing, malformed, invalid, or expired API key
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: API key lacks the scope required by this operation
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Hedge intent not found
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
            Retry-After:
              description: Seconds until the client should retry.
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Rate limit store or hedge status store unavailable
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /settlements/{settlementEventId}:
    get:
      operationId: getSettlementEvent
      summary: Get settlement event status by id
      x-required-scope: status:read
      description: >
        Returns the accepted settlement event consumed by the backend to update
        inventory. Receipt-confirmed deployments derive these records from
        RFQSettlement events using a chain id, transaction hash, and log index
        idempotency key, persist the emitted nonce for chainId/user/nonce
        reconciliation, and persist the event block number for reorg-aware
        reconciliation. Explicit local simulation produces the same record shape
        with synthetic chain evidence.
      tags:
        - Settlement
      parameters:
        - name: settlementEventId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: "^[A-Za-z0-9_:-]+$"
          description: Settlement event identifier returned by POST /submit. The gateway accepts only 1 to 128 letters, numbers, underscore, colon, or hyphen characters.
      responses:
        "200":
          description: Settlement event status
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettlementEventStatus"
        "401":
          description: Missing, malformed, invalid, or expired API key
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: API key lacks the scope required by this operation
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Settlement event not found
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
            Retry-After:
              description: Seconds until the client should retry.
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Rate limit store or settlement event store unavailable
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /pnl:
    get:
      operationId: getPnlSummary
      summary: Get realized PnL summary
      x-required-scope: pnl:read
      description: >
        Returns gross settlement PnL records valued against each quote's persisted
        market snapshot. The model normalizes token decimals, reports tokenOut base
        units, groups totals by chain and output token, and excludes fees, gas, and
        hedge execution.
      tags:
        - PnL
      responses:
        "200":
          description: Realized PnL summary
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PnlSummary"
        "401":
          description: Missing, malformed, invalid, or expired API key
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: API key lacks the scope required by this operation
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
            Retry-After:
              description: Seconds until the client should retry.
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Rate limit store or PnL store unavailable
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Scoped key in keyId.secret form. Store only SHA-256(secret) in RFQ_API_KEY_CONFIG_JSON and rotate with overlapping key ids.
  headers:
    TraceId:
      description: Request correlation id attached to every response. ErrorResponse.traceId contains the same value for structured error bodies. Safe tr_-prefixed request header values may be echoed; unsafe values are replaced with a generated request id.
      schema:
        type: string
        pattern: "^tr_.+"
  schemas:
...
    QuoteRequest:
      type: object
      additionalProperties: false
      required:
        - chainId
        - user
        - tokenIn
        - tokenOut
        - amountIn
        - slippageBps
      properties:
        chainId:
          type: integer
          minimum: 1
          maximum: 9007199254740991
          description: Positive JavaScript safe integer chain id.
        user:
          $ref: "#/components/schemas/Address"
        tokenIn:
          $ref: "#/components/schemas/Address"
        tokenOut:
          $ref: "#/components/schemas/Address"
        amountIn:
          $ref: "#/components/schemas/PositiveUIntString"
        slippageBps:
          type: integer
          minimum: 0
          maximum: 10000
    QuoteResponse:
      type: object
      additionalProperties: false
      required:
        - quoteId
        - snapshotId
        - amountOut
        - minAmountOut
        - deadline
        - nonce
        - signature
      properties:
        quoteId:
          $ref: "#/components/schemas/SafeIdentifier"
        snapshotId:
          $ref: "#/components/schemas/SafeIdentifier"
        amountOut:
          $ref: "#/components/schemas/PositiveUIntString"
        minAmountOut:
          $ref: "#/components/schemas/PositiveUIntString"
        deadline:
          type: integer
          description: Unix timestamp generated from the configured RFQ_QUOTE_TTL_SECONDS signed quote lifetime.
        nonce:
          $ref: "#/components/schemas/PositiveUIntString"
        signature:
          type: string
          pattern: "^0x[a-fA-F0-9]{130}$"
          minLength: 132
          maxLength: 132
          description: 65-byte canonical low-s EIP-712 signature encoded as 0x-prefixed hex.
...